Move StimulusReflex::Channel to app/ and allow for a configurable parent channel
