+++ lib/generators/stimulus_reflex/templates/config/initializers/stimulus_reflex.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 StimulusReflex.configure do |config|
-  # configuration options explained here:
-  # Short description of what the config variable does, best with the available options.
-  # config.example = true
+  # Override the parent class that the StimulusReflex ActionCable channel inherits from
+  # config.parent_channel = "ApplicationCable::Channel"
 end
