+++ lib/stimulus_reflex.rb
@@ -12,7 +12,6 @@ require "stimulus_reflex/version"
 require "stimulus_reflex/configuration"
 require "stimulus_reflex/reflex"
 require "stimulus_reflex/element"
-require "stimulus_reflex/channel"
 require "stimulus_reflex/sanity_checker"
 require "stimulus_reflex/broadcasters/broadcaster"
 require "stimulus_reflex/broadcasters/nothing_broadcaster"
